wincng: prefer `ULONG`/`DWORD` over `unsigned long`
To match with the types used by the `Crypt*()` (uses `DWORD`) and
`BCrypt*()` (uses `ULONG`) Windows APIs.

This patch doesn't change data width or signedness.

Closes #1165

+++ src/wincng.h
@@ -144,8 +144,8 @@ void _libssh2_wincng_free(void);
 typedef struct __libssh2_wincng_hash_ctx {
     BCRYPT_HASH_HANDLE hHash;
     unsigned char *pbHashObject;
-    unsigned long dwHashObject;
-    unsigned long cbHash;
+    ULONG dwHashObject;
+    ULONG cbHash;
 } _libssh2_wincng_hash_ctx;
 
 /*
@@ -159,7 +159,7 @@ typedef struct __libssh2_wincng_hash_ctx {
 #define libssh2_sha1_update(ctx, data, datalen) \
     _libssh2_wincng_hash_update(&ctx, \
                                 (const unsigned char *) data, \
-                                (unsigned long) datalen)
+                                (ULONG) datalen)
 #define libssh2_sha1_final(ctx, hash) \
     _libssh2_wincng_hash_final(&ctx, hash)
 #define libssh2_sha1(data, datalen, hash) \
@@ -173,7 +173,7 @@ typedef struct __libssh2_wincng_hash_ctx {
 #define libssh2_sha256_update(ctx, data, datalen) \
     _libssh2_wincng_hash_update(&ctx, \
                                 (const unsigned char *) data, \
-                                (unsigned long) datalen)
+                                (ULONG) datalen)
 #define libssh2_sha256_final(ctx, hash) \
     _libssh2_wincng_hash_final(&ctx, hash)
 #define libssh2_sha256(data, datalen, hash) \
@@ -187,7 +187,7 @@ typedef struct __libssh2_wincng_hash_ctx {
 #define libssh2_sha384_update(ctx, data, datalen) \
     _libssh2_wincng_hash_update(&ctx, \
                                 (const unsigned char *) data, \
-                                (unsigned long) datalen)
+                                (ULONG) datalen)
 #define libssh2_sha384_final(ctx, hash) \
     _libssh2_wincng_hash_final(&ctx, hash)
 #define libssh2_sha384(data, datalen, hash) \
@@ -201,7 +201,7 @@ typedef struct __libssh2_wincng_hash_ctx {
 #define libssh2_sha512_update(ctx, data, datalen) \
     _libssh2_wincng_hash_update(&ctx, \
                                 (const unsigned char *) data, \
-                                (unsigned long) datalen)
+                                (ULONG) datalen)
 #define libssh2_sha512_final(ctx, hash) \
     _libssh2_wincng_hash_final(&ctx, hash)
 #define libssh2_sha512(data, datalen, hash) \
@@ -215,7 +215,7 @@ typedef struct __libssh2_wincng_hash_ctx {
 #define libssh2_md5_update(ctx, data, datalen) \
     _libssh2_wincng_hash_update(&ctx, \
                                 (const unsigned char *) data, \
-                                (unsigned long) datalen)
+                                (ULONG) datalen)
 #define libssh2_md5_final(ctx, hash) \
     _libssh2_wincng_hash_final(&ctx, hash)
 #define libssh2_md5(data, datalen, hash) \
@@ -231,25 +231,25 @@ typedef struct __libssh2_wincng_hash_ctx {
 #define libssh2_hmac_sha1_init(ctx, key, keylen) \
     _libssh2_wincng_hash_init(ctx, _libssh2_wincng.hAlgHmacSHA1, \
                               SHA_DIGEST_LENGTH, \
-                              key, (unsigned long) keylen)
+                              key, (ULONG) keylen)
 #define libssh2_hmac_md5_init(ctx, key, keylen) \
     _libssh2_wincng_hash_init(ctx, _libssh2_wincng.hAlgHmacMD5, \
                               MD5_DIGEST_LENGTH, \
-                              key, (unsigned long) keylen)
+                              key, (ULONG) keylen)
 #define libssh2_hmac_ripemd160_init(ctx, key, keylen)
     /* not implemented */
 #define libssh2_hmac_sha256_init(ctx, key, keylen) \
     _libssh2_wincng_hash_init(ctx, _libssh2_wincng.hAlgHmacSHA256, \
                               SHA256_DIGEST_LENGTH, \
-                              key, (unsigned long) keylen)
+                              key, (ULONG) keylen)
 #define libssh2_hmac_sha512_init(ctx, key, keylen) \
     _libssh2_wincng_hash_init(ctx, _libssh2_wincng.hAlgHmacSHA512, \
                               SHA512_DIGEST_LENGTH, \
-                              key, (unsigned long) keylen)
+                              key, (ULONG) keylen)
 #define libssh2_hmac_update(ctx, data, datalen) \
     _libssh2_wincng_hash_update(&ctx, \
                                 (const unsigned char *) data, \
-                                (unsigned long) datalen)
+                                (ULONG) datalen)
 #define libssh2_hmac_final(ctx, hash) \
     _libssh2_wincng_hmac_final(&ctx, hash)
 #define libssh2_hmac_cleanup(ctx) \
@@ -264,7 +264,7 @@ typedef struct __libssh2_wincng_hash_ctx {
 typedef struct __libssh2_wincng_key_ctx {
     BCRYPT_KEY_HANDLE hKey;
     void *pbKeyObject;
-    unsigned long cbKeyObject;
+    DWORD cbKeyObject;
 } _libssh2_wincng_key_ctx;
 
 
@@ -343,10 +343,10 @@ struct _libssh2_wincng_cipher_ctx {
     unsigned char *pbKeyObject;
     unsigned char *pbIV;
     unsigned char *pbCtr;
-    unsigned long dwKeyObject;
-    unsigned long dwIV;
-    unsigned long dwBlockLength;
-    unsigned long dwCtrLength;
+    ULONG dwKeyObject;
+    ULONG dwIV;
+    ULONG dwBlockLength;
+    ULONG dwCtrLength;
 };
 
 #define _libssh2_cipher_ctx struct _libssh2_wincng_cipher_ctx
@@ -357,7 +357,7 @@ struct _libssh2_wincng_cipher_ctx {
 
 struct _libssh2_wincng_cipher_type {
     BCRYPT_ALG_HANDLE *phAlg;
-    unsigned long dwKeyLength;
+    ULONG dwKeyLength;
     int useIV;      /* TODO: Convert to bool when a C89 compatible bool type
                        is defined */
     int ctrMode;
@@ -402,7 +402,7 @@ struct _libssh2_wincng_cipher_type {
 
 struct _libssh2_wincng_bignum {
     unsigned char *bignum;
-    unsigned long length;
+    ULONG length;
 };
 
 #define _libssh2_bn struct _libssh2_wincng_bignum
@@ -420,7 +420,7 @@ _libssh2_bn *_libssh2_wincng_bignum_init(void);
 #define _libssh2_bn_set_word(bn, word) \
     _libssh2_wincng_bignum_set_word(bn, word)
 #define _libssh2_bn_from_bin(bn, len, bin) \
-    _libssh2_wincng_bignum_from_bin(bn, (unsigned long) len, bin)
+    _libssh2_wincng_bignum_from_bin(bn, (ULONG) len, bin)
 #define _libssh2_bn_to_bin(bn, bin) \
     _libssh2_wincng_bignum_to_bin(bn, bin)
 #define _libssh2_bn_bytes(bn) bn->length
@@ -469,18 +469,18 @@ int _libssh2_wincng_random(void *buf, size_t len);
 
 int
 _libssh2_wincng_hash_init(_libssh2_wincng_hash_ctx *ctx,
-                          BCRYPT_ALG_HANDLE hAlg, unsigned long hashlen,
-                          unsigned char *key, unsigned long keylen);
+                          BCRYPT_ALG_HANDLE hAlg, ULONG hashlen,
+                          unsigned char *key, ULONG keylen);
 int
 _libssh2_wincng_hash_update(_libssh2_wincng_hash_ctx *ctx,
-                            const unsigned char *data, unsigned long datalen);
+                            const unsigned char *data, ULONG datalen);
 int
 _libssh2_wincng_hash_final(_libssh2_wincng_hash_ctx *ctx,
                            unsigned char *hash);
 int
-_libssh2_wincng_hash(const unsigned char *data, unsigned long datalen,
+_libssh2_wincng_hash(const unsigned char *data, ULONG datalen,
                      BCRYPT_ALG_HANDLE hAlg,
-                     unsigned char *hash, unsigned long hashlen);
+                     unsigned char *hash, ULONG hashlen);
 
 int
 _libssh2_wincng_hmac_final(_libssh2_wincng_hash_ctx *ctx,
@@ -607,11 +607,11 @@ _libssh2_wincng_cipher_dtor(_libssh2_cipher_ctx *ctx);
 _libssh2_bn *
 _libssh2_wincng_bignum_init(void);
 int
-_libssh2_wincng_bignum_set_word(_libssh2_bn *bn, unsigned long word);
-unsigned long
+_libssh2_wincng_bignum_set_word(_libssh2_bn *bn, ULONG word);
+ULONG
 _libssh2_wincng_bignum_bits(const _libssh2_bn *bn);
 void
-_libssh2_wincng_bignum_from_bin(_libssh2_bn *bn, unsigned long len,
+_libssh2_wincng_bignum_from_bin(_libssh2_bn *bn, ULONG len,
                                 const unsigned char *bin);
 void
 _libssh2_wincng_bignum_to_bin(const _libssh2_bn *bn, unsigned char *bin);
